handled a bunch of edge cases mostly related to initial state.
for instance what happens when the web part hasn't been configured with a list title yet? what happens when the user resets the list title to blank? what happens when the list is empty? what happens when the user deletes the last image? all of these cases caused errors, and now they don't.

also cleaned up the HTML and CSS a bit for better presentation.